Fix for #1033 : Refactor PredictionTask.java for increased reusability
