+++ inception-active-learning/src/main/java/de/tudarmstadt/ukp/inception/active/learning/sidebar/ActiveLearningSidebar.java
@@ -123,7 +123,7 @@ import de.tudarmstadt.ukp.inception.recommendation.api.model.SuggestionGroup.Del
 import de.tudarmstadt.ukp.inception.recommendation.event.AjaxPredictionsSwitchedEvent;
 import de.tudarmstadt.ukp.inception.recommendation.event.AjaxRecommendationAcceptedEvent;
 import de.tudarmstadt.ukp.inception.recommendation.event.AjaxRecommendationRejectedEvent;
-import de.tudarmstadt.ukp.inception.recommendation.tasks.PredictionTask;
+import de.tudarmstadt.ukp.inception.recommendation.util.PredictionUtil;
 
 public class ActiveLearningSidebar
     extends AnnotationSidebar_ImplBase
@@ -184,7 +184,7 @@ public class ActiveLearningSidebar
     private VID highlightVID;
     private Offset highlightSpan;
     private boolean protectHighlight;
-    
+
     public ActiveLearningSidebar(String aId, IModel<AnnotatorState> aModel,
             AnnotationActionHandler aActionHandler, JCasProvider aJCasProvider,
             AnnotationPage aAnnotationPage)
@@ -195,9 +195,9 @@ public class ActiveLearningSidebar
 
         // Instead of maintaining the AL state in the sidebar, we maintain it in the page because
         // that way we persists even if we switch to another sidebar tab
-        alStateModel = new CompoundPropertyModel<>(LambdaModelAdapter.of(
-            () -> aAnnotationPage.getMetaData(CURRENT_AL_USER_STATE),
-            alState -> aAnnotationPage.setMetaData(CURRENT_AL_USER_STATE, alState)));
+        alStateModel = new CompoundPropertyModel<>(
+                LambdaModelAdapter.of(() -> aAnnotationPage.getMetaData(CURRENT_AL_USER_STATE),
+                        alState -> aAnnotationPage.setMetaData(CURRENT_AL_USER_STATE, alState)));
 
         // Set up the AL state in the page if it is not already there or if for some reason the
         // suggestions have completely disappeared (e.g. after a system restart)
@@ -207,9 +207,10 @@ public class ActiveLearningSidebar
         if (aAnnotationPage.getMetaData(CURRENT_AL_USER_STATE) == null || model == null) {
             ActiveLearningUserState alState = new ActiveLearningUserState();
             alState.setStrategy(new UncertaintySamplingStrategy());
-            alStateModel.setObject(alState);;
+            alStateModel.setObject(alState);
+            ;
         }
-        
+
         mainContainer = new WebMarkupContainer(CID_MAIN_CONTAINER);
         mainContainer.setOutputMarkupId(true);
         mainContainer.add(createNoRecommendersMessage());
@@ -219,7 +220,7 @@ public class ActiveLearningSidebar
         mainContainer.add(createRecommendationOperationForm());
         mainContainer.add(createLearningHistory());
         add(mainContainer);
-        
+
         confirmationDialog = new ConfirmationDialog(CID_CONFIRMATION_DIALOG);
         confirmationDialog.setAutoSize(true);
         add(confirmationDialog);
@@ -232,8 +233,7 @@ public class ActiveLearningSidebar
             // default choice in the active learning mode.
             List<AnnotationLayer> layersWithRecommenders = listLayersWithRecommenders();
             if (layersWithRecommenders.contains(getModelObject().getDefaultAnnotationLayer())) {
-                alStateModel.getObject()
-                    .setLayer(getModelObject().getDefaultAnnotationLayer());
+                alStateModel.getObject().setLayer(getModelObject().getDefaultAnnotationLayer());
             }
             // If the currently selected layer has no recommenders, use the first one which has
             else if (!layersWithRecommenders.isEmpty()) {
@@ -247,10 +247,10 @@ public class ActiveLearningSidebar
             }
         }
         Label noRecommendersMessage = new Label(CID_NO_RECOMMENDERS, "None of the layers have any "
-            + "recommenders configured. Please set the recommenders first in the Project "
-            + "Settings.");
-        noRecommendersMessage.add(LambdaBehavior.onConfigure(component -> component.setVisible
-            (!alStateModel.getObject().isDoExistRecommenders())));
+                + "recommenders configured. Please set the recommenders first in the Project "
+                + "Settings.");
+        noRecommendersMessage.add(LambdaBehavior.onConfigure(component -> component
+                .setVisible(!alStateModel.getObject().isDoExistRecommenders())));
         return noRecommendersMessage;
     }
 
@@ -262,16 +262,16 @@ public class ActiveLearningSidebar
         layersDropdown.setModel(alStateModel.bind("layer"));
         layersDropdown.setChoices(LoadableDetachableModel.of(this::listLayersWithRecommenders));
         layersDropdown.setChoiceRenderer(new LambdaChoiceRenderer<>(AnnotationLayer::getUiName));
-        layersDropdown.add(LambdaBehavior.onConfigure(it -> it.setEnabled(!alStateModel
-            .getObject().isSessionActive())));
+        layersDropdown.add(LambdaBehavior
+                .onConfigure(it -> it.setEnabled(!alStateModel.getObject().isSessionActive())));
         layersDropdown.setOutputMarkupId(true);
         layersDropdown.setRequired(true);
         form.add(layersDropdown);
-        
+
         form.add(new LambdaAjaxSubmitLink(CID_START_SESSION_BUTTON, this::actionStartSession)
                 .add(visibleWhen(() -> !alStateModel.getObject().isSessionActive())));
         form.add(new LambdaAjaxLink(CID_STOP_SESSION_BUTTON, this::actionStopSession)
-            .add(visibleWhen(() -> alStateModel.getObject().isSessionActive())));
+                .add(visibleWhen(() -> alStateModel.getObject().isSessionActive())));
         form.add(visibleWhen(() -> alStateModel.getObject().isDoExistRecommenders()));
 
         return form;
@@ -282,12 +282,12 @@ public class ActiveLearningSidebar
         return recommendationService
                 .listLayersWithEnabledRecommenders(getModelObject().getProject());
     }
-    
+
     private void actionStartSession(AjaxRequestTarget target, Form<?> form)
     {
         ActiveLearningUserState alState = alStateModel.getObject();
         AnnotatorState state = getModelObject();
-        
+
         // Start new session
         alState.setSessionActive(true);
 
@@ -298,7 +298,7 @@ public class ActiveLearningSidebar
         applicationEventPublisherHolder.get().publishEvent(new ActiveLearningSessionStartedEvent(
                 this, state.getProject(), state.getUser().getUsername()));
     }
-    
+
     private void actionStopSession(AjaxRequestTarget target)
     {
         ActiveLearningUserState alState = alStateModel.getObject();
@@ -320,13 +320,13 @@ public class ActiveLearningSidebar
             protectHighlight = false;
             return;
         }
-        
+
         LOG.trace("Active learning sidebar set highlight suggestion: {}", aSuggestion);
         highlightVID = aSuggestion.getVID();
         highlightSpan = new Offset(aSuggestion.getBegin(), aSuggestion.getEnd());
         highlightDocumentName = aSuggestion.getDocumentName();
     }
-    
+
     private void setHighlight(LearningRecord aRecord)
     {
         LOG.trace("Active learning sidebar set highlight history record: {}", aRecord);
@@ -338,20 +338,19 @@ public class ActiveLearningSidebar
         // - use removes an ACCEPT history item
         // - user clicks then on another history item
         // - ... but during the subsequent rendering the "moveToNextSuggestion" method sets or
-        //   clears the highlight.
+        // clears the highlight.
         protectHighlight = true;
     }
-    
+
     private void setHighlight(SourceDocument aDocument, AnnotationFS aAnnotation)
     {
         LOG.trace("Active learning sidebar set highlight annotation: {}", aAnnotation);
         highlightVID = new VID(WebAnnoCasUtil.getAddr(aAnnotation));
-        highlightSpan = new Offset(aAnnotation.getBegin(),
-                aAnnotation.getEnd());
+        highlightSpan = new Offset(aAnnotation.getBegin(), aAnnotation.getEnd());
         highlightDocumentName = aDocument.getName();
         protectHighlight = false;
     }
-    
+
     private void clearHighlight()
     {
         if (protectHighlight) {
@@ -359,7 +358,7 @@ public class ActiveLearningSidebar
             protectHighlight = false;
             return;
         }
-        
+
         LOG.trace("Active learning sidebar cleared highlights");
         highlightDocumentName = null;
         highlightSpan = null;
@@ -372,10 +371,9 @@ public class ActiveLearningSidebar
                 "There are no further suggestions.");
         noRecommendation.add(visibleWhen(() -> {
             ActiveLearningUserState alState = alStateModel.getObject();
-            return alState.isSessionActive() 
-                    && !alState.getSuggestion().isPresent()
-                    && !activeLearningService.hasSkippedSuggestions(
-                            getModelObject().getUser(), alState.getLayer());
+            return alState.isSessionActive() && !alState.getSuggestion().isPresent()
+                    && !activeLearningService.hasSkippedSuggestions(getModelObject().getUser(),
+                            alState.getLayer());
         }));
         noRecommendation.setOutputMarkupPlaceholderTag(true);
         return noRecommendation;
@@ -386,10 +384,9 @@ public class ActiveLearningSidebar
         Form<Void> form = new Form<>(CID_LEARN_FROM_SKIPPED_RECOMMENDATION_FORM);
         form.add(LambdaBehavior.visibleWhen(() -> {
             ActiveLearningUserState alState = alStateModel.getObject();
-            return alState.isSessionActive() 
-                    && !alState.getSuggestion().isPresent()
-                    && activeLearningService.hasSkippedSuggestions(
-                            getModelObject().getUser(), alState.getLayer());
+            return alState.isSessionActive() && !alState.getSuggestion().isPresent()
+                    && activeLearningService.hasSkippedSuggestions(getModelObject().getUser(),
+                            alState.getLayer());
         }));
         form.setOutputMarkupPlaceholderTag(true);
         form.add(new Label(CID_ONLY_SKIPPED_RECOMMENDATION_LABEL,
@@ -410,13 +407,12 @@ public class ActiveLearningSidebar
         // The history records caused suggestions to disappear. Since visibility is only fully
         // recalculated when new predictions come in, we need to update the visibility explicitly
         // here
-        alState.getSuggestions().stream()
-            .flatMap(group -> group.stream())
-            .forEach(suggestion -> suggestion.show(FLAG_SKIPPED));
-        
+        alState.getSuggestions().stream().flatMap(group -> group.stream())
+                .forEach(suggestion -> suggestion.show(FLAG_SKIPPED));
+
         refreshSuggestions();
         moveToNextRecommendation(aTarget, false);
-        
+
         aTarget.add(mainContainer);
     }
 
@@ -430,19 +426,16 @@ public class ActiveLearningSidebar
         recommendationForm.setOutputMarkupPlaceholderTag(true);
 
         recommendationForm.add(createJumpToSuggestionLink());
-        recommendationForm.add(new Label(CID_RECOMMENDED_PREDITION,
-                LoadableDetachableModel.of(() -> alStateModel.getObject()
-                        .getSuggestion().map(this::formatLabel).orElse(null))));
-        recommendationForm.add(new Label(CID_RECOMMENDED_CONFIDENCE, () -> 
-                alStateModel.getObject().getSuggestion()
-                        .map(AnnotationSuggestion::getConfidence).orElse(null)));
-        recommendationForm.add(new Label(CID_RECOMMENDED_DIFFERENCE, () -> 
-                alStateModel.getObject().getCurrentDifference()
-                        .map(Delta::getDelta).orElse(null)));
+        recommendationForm.add(
+                new Label(CID_RECOMMENDED_PREDITION, LoadableDetachableModel.of(() -> alStateModel
+                        .getObject().getSuggestion().map(this::formatLabel).orElse(null))));
+        recommendationForm.add(new Label(CID_RECOMMENDED_CONFIDENCE, () -> alStateModel.getObject()
+                .getSuggestion().map(AnnotationSuggestion::getConfidence).orElse(null)));
+        recommendationForm.add(new Label(CID_RECOMMENDED_DIFFERENCE, () -> alStateModel.getObject()
+                .getCurrentDifference().map(Delta::getDelta).orElse(null)));
         recommendationForm.add((alStateModel.getObject().getLayer() != null
-            && alStateModel.getObject().getSuggestion().isPresent()) ?
-            initializeFeatureEditor() :
-            new Label(CID_EDITOR).setVisible(false));
+                && alStateModel.getObject().getSuggestion().isPresent()) ? initializeFeatureEditor()
+                        : new Label(CID_EDITOR).setVisible(false));
 
         recommendationForm.add(new LambdaAjaxButton<>(CID_ANNOTATE_BUTTON, this::actionAnnotate));
         recommendationForm.add(new LambdaAjaxLink(CID_SKIP_BUTTON, this::actionSkip));
@@ -474,11 +467,10 @@ public class ActiveLearningSidebar
         return link;
     }
 
-    private void actionJumpToSuggestion(AjaxRequestTarget aTarget)
-        throws IOException
+    private void actionJumpToSuggestion(AjaxRequestTarget aTarget) throws IOException
     {
         ActiveLearningUserState alState = alStateModel.getObject();
-        
+
         AnnotationSuggestion suggestion = alState.getSuggestion().get();
 
         if (LOG.isDebugEnabled()) {
@@ -511,24 +503,24 @@ public class ActiveLearningSidebar
         recommendationForm.addOrReplace(editor);
         aTarget.add(mainContainer);
     }
-    
+
     private FeatureEditor createFeatureEditor(AnnotationSuggestion aCurrentRecommendation)
     {
         AnnotatorState state = ActiveLearningSidebar.this.getModelObject();
         ActiveLearningUserState alState = alStateModel.getObject();
-        
+
         // Obtain the feature state which serves as a model to the editor
         AnnotationFeature feat = annotationService.getFeature(aCurrentRecommendation.getFeature(),
                 alState.getLayer());
         FeatureSupport<?> featureSupport = featureSupportRegistry.getFeatureSupport(feat);
-        // We get away with passing "null" here instead of the CAS because we currently 
+        // We get away with passing "null" here instead of the CAS because we currently
         // have no recommenders for any feature types that actually need the CAS (i.e.
         // link feature types and the likes).
         Object wrappedFeatureValue = featureSupport.wrapFeatureValue(feat, null,
                 aCurrentRecommendation.getLabel());
         FeatureState featureState = new FeatureState(feat, (Serializable) wrappedFeatureValue);
-        
-        // Populate the tagset moving the tags with recommended labels to the top 
+
+        // Populate the tagset moving the tags with recommended labels to the top
         List<Tag> tagList = annotationService.listTags(feat.getTagset());
         List<Tag> reorderedTagList = new ArrayList<>();
         if (tagList.size() > 0) {
@@ -541,9 +533,8 @@ public class ActiveLearningSidebar
                     aCurrentRecommendation.getFeature());
             // get all the label of the predictions (e.g. "NN")
             List<String> allRecommendationLabels = allRecommendations.stream()
-                    .map(ao -> ao.getLabel())
-                    .collect(Collectors.toList());
-            
+                    .map(ao -> ao.getLabel()).collect(Collectors.toList());
+
             for (Tag tag : tagList) {
                 // add the tags which contain the prediction-labels to the beginning of a tagset
                 if (allRecommendationLabels.contains(tag.getName())) {
@@ -551,15 +542,15 @@ public class ActiveLearningSidebar
                     reorderedTagList.add(tag);
                 }
             }
-            
+
             // remove these tags containing the prediction-labels
             tagList.removeAll(reorderedTagList);
-            
+
             // add the rest tags to the tagset after these
             reorderedTagList.addAll(tagList);
         }
         featureState.tagset = reorderedTagList;
-        
+
         // Finally, create the editor
         FeatureEditor featureEditor = featureSupport.createEditor(CID_EDITOR, mainContainer,
                 this.getActionHandler(), this.getModel(), Model.of(featureState));
@@ -568,7 +559,7 @@ public class ActiveLearningSidebar
                 && alState.getSuggestion().isPresent()));
         return featureEditor;
     }
-    
+
     private void writeLearningRecordInDatabaseAndEventLog(
             AnnotationSuggestion aCurrentRecommendation, LearningRecordType aUserAction)
     {
@@ -588,8 +579,8 @@ public class ActiveLearningSidebar
                 aSuggestion.getDocumentName());
 
         // Log the action to the learning record
-        learningRecordService.logRecord(sourceDoc, state.getUser().getUsername(),
-                aSuggestion, aAnnotationValue, alState.getLayer(), feat, aUserAction,
+        learningRecordService.logRecord(sourceDoc, state.getUser().getUsername(), aSuggestion,
+                aAnnotationValue, alState.getLayer(), feat, aUserAction,
                 LearningRecordChangeLocation.AL_SIDEBAR);
 
         // Send an application event that the suggestion has been rejected
@@ -606,27 +597,26 @@ public class ActiveLearningSidebar
 
     /**
      * Accept a suggestion or a corrected suggestion via the sidebar. If the value in the feature
-     * editor corresponds to the originally suggested label, an acceptance is logged, otherwise
-     * a correction is logged.
+     * editor corresponds to the originally suggested label, an acceptance is logged, otherwise a
+     * correction is logged.
      * 
      * <ul>
-     * <li>Creates a new annotation or updates an existing one with a new feature
-     * value.</li>
+     * <li>Creates a new annotation or updates an existing one with a new feature value.</li>
      * <li>Marks the suggestions as hidden (not visible).</li>
      * <li>Logs the accepting to the learning log.</li>
      * <li>Sends events to the UI and application informing other components about the action.</li>
      * </ul>
-     */    
+     */
     private void actionAnnotate(AjaxRequestTarget aTarget, Form<Void> aForm)
         throws IOException, AnnotationException
     {
         AnnotatorState state = ActiveLearningSidebar.this.getModelObject();
         ActiveLearningUserState alState = alStateModel.getObject();
-        
+
         // There is always a current recommendation when we get here because if there is none, the
         // button to accept the recommendation is not visible.
         AnnotationSuggestion suggestion = alState.getSuggestion().get();
-        
+
         // Create AnnotationFeature and FeatureSupport
         AnnotationFeature feat = annotationService.getFeature(suggestion.getFeature(),
                 alState.getLayer());
@@ -647,10 +637,10 @@ public class ActiveLearningSidebar
         AnnotationFeature feature = annotationService.getFeature(suggestion.getFeature(), layer);
         recommendationService.upsertFeature(annotationService, sourceDoc, username, jCas, layer,
                 feature, selectedValue, suggestion.getBegin(), suggestion.getEnd());
-        
+
         // Save CAS after annotation has been created
         documentService.writeAnnotationCas(jCas, sourceDoc, state.getUser(), true);
-        
+
         // If the currently displayed document is the same one where the annotation was created,
         // then update timestamp in state to avoid concurrent modification errors
         if (Objects.equals(state.getDocument().getId(), sourceDoc.getId())) {
@@ -660,14 +650,14 @@ public class ActiveLearningSidebar
                 state.setAnnotationDocumentTimestamp(diskTimestamp.get());
             }
         }
-        
+
         suggestion.hide(selectedValue.equals(suggestion.getLabel()) ? FLAG_TRANSIENT_ACCEPTED
                 : FLAG_TRANSIENT_CORRECTED);
 
         // Log the action to the learning record
-        writeLearningRecordInDatabaseAndEventLog(suggestion, 
+        writeLearningRecordInDatabaseAndEventLog(suggestion,
                 selectedValue.equals(suggestion.getLabel()) ? ACCEPTED : CORRECTED, selectedValue);
-        
+
         recommendationService.getPredictions(state.getUser(), state.getProject())
                 .getPredictionsByTokenAndFeature(suggestion.getDocumentName(),
                         alStateModel.getObject().getLayer(), suggestion.getBegin(),
@@ -691,13 +681,13 @@ public class ActiveLearningSidebar
             moveToNextRecommendation(aTarget, false);
         });
     }
-    
+
     private void moveToNextRecommendation(IPartialPageRequestHandler aTarget, boolean aStay)
     {
         AnnotatorState state = getModelObject();
         ActiveLearningUserState alState = alStateModel.getObject();
-        
-        // Clear the annotation detail editor and the selection to avoid confusions with the 
+
+        // Clear the annotation detail editor and the selection to avoid confusions with the
         // highlight because the selection highlight from the right sidebar and the one from the
         // AL sidebar have the same color!
         if (!aStay) {
@@ -711,7 +701,7 @@ public class ActiveLearningSidebar
                 alState.getSuggestions());
         Optional<AnnotationSuggestion> prevSuggestion = alState.getSuggestion();
         alState.setCurrentDifference(recommendationDifference);
-        
+
         // If the active suggestion has changed, inform the user
         if (prevSuggestion.isPresent() && (alState.getSuggestion().isPresent()
                 && !alState.getSuggestion().get().equals(prevSuggestion.get()))) {
@@ -738,7 +728,7 @@ public class ActiveLearningSidebar
             clearHighlight();
             aTarget.add(mainContainer);
             if (!aStay) {
-                //Main editor
+                // Main editor
                 annotationPage.actionRefreshDocument((AjaxRequestTarget) aTarget);
             }
             return;
@@ -749,10 +739,10 @@ public class ActiveLearningSidebar
             AnnotationSuggestion suggestion = alState.getSuggestion().get();
             SourceDocument sourceDocument = documentService.getSourceDocument(state.getProject(),
                     suggestion.getDocumentName());
-            
+
             // Refresh feature editor
             refreshFeatureEditor(aTarget, suggestion);
-            
+
             if (!aStay) {
                 // Open the corresponding document in the main editor and jump to the respective
                 // location
@@ -760,12 +750,12 @@ public class ActiveLearningSidebar
                         suggestion.getBegin(), suggestion.getEnd());
             }
             setHighlight(suggestion);
-            
+
             if (!aStay) {
-                //Main editor
+                // Main editor
                 annotationPage.actionRefreshDocument((AjaxRequestTarget) aTarget);
             }
-            
+
             // Obtain some left and right context of the active suggestion while we have easy
             // access to the document which contains the current suggestion
             JCas cas;
@@ -781,7 +771,7 @@ public class ActiveLearningSidebar
                     text.substring(Math.max(0, suggestion.getBegin() - 20), suggestion.getBegin()));
             alState.setRightContext(text.substring(suggestion.getEnd(),
                     Math.min(suggestion.getEnd() + 20, text.length())));
-            
+
             // Send an application event that the suggestion has been rejected
             List<AnnotationSuggestion> alternativeSuggestions = recommendationService
                     .getPredictions(state.getUser(), state.getProject())
@@ -800,15 +790,15 @@ public class ActiveLearningSidebar
             aTarget.addChildren(getPage(), IFeedback.class);
         }
     }
-    
+
     private Form<?> createLearningHistory()
     {
         Form<?> learningHistoryForm = new Form<Void>(CID_LEARNING_HISTORY_FORM)
         {
             private static final long serialVersionUID = -961690443085882064L;
         };
-        learningHistoryForm.add(LambdaBehavior.onConfigure(component -> component
-            .setVisible(alStateModel.getObject().isSessionActive())));
+        learningHistoryForm.add(LambdaBehavior.onConfigure(
+                component -> component.setVisible(alStateModel.getObject().isSessionActive())));
         learningHistoryForm.setOutputMarkupPlaceholderTag(true);
         learningHistoryForm.setOutputMarkupId(true);
 
@@ -831,22 +821,22 @@ public class ActiveLearningSidebar
                 String recFeatureValue;
                 if (recAnnotationFeature != null) {
                     FeatureSupport featureSupport = featureSupportRegistry
-                        .getFeatureSupport(recAnnotationFeature);
-                    recFeatureValue = featureSupport
-                        .renderFeatureValue(recAnnotationFeature, rec.getAnnotation());
+                            .getFeatureSupport(recAnnotationFeature);
+                    recFeatureValue = featureSupport.renderFeatureValue(recAnnotationFeature,
+                            rec.getAnnotation());
                 }
                 else {
                     recFeatureValue = rec.getAnnotation();
                 }
-                LambdaAjaxLink textLink = new LambdaAjaxLink(CID_JUMP_TO_ANNOTATION, _target -> 
-                        actionSelectHistoryItem(_target, item.getModelObject()));
+                LambdaAjaxLink textLink = new LambdaAjaxLink(CID_JUMP_TO_ANNOTATION,
+                        _target -> actionSelectHistoryItem(_target, item.getModelObject()));
                 textLink.setBody(rec::getTokenText);
                 item.add(textLink);
 
                 item.add(new Label(CID_RECOMMENDED_ANNOTATION, recFeatureValue));
                 item.add(new Label(CID_USER_ACTION, rec.getUserAction()));
-                item.add(
-                    new LambdaAjaxLink(CID_REMOVE_RECORD, t -> actionRemoveHistoryItem(t, rec)));
+                item.add(new LambdaAjaxLink(CID_REMOVE_RECORD,
+                        t -> actionRemoveHistoryItem(t, rec)));
             }
         };
         learningRecords = LambdaModel.of(this::listLearningRecords);
@@ -860,7 +850,7 @@ public class ActiveLearningSidebar
      * <ul>
      * <li>the main editor should jump to the location of the history item</li>
      * <li>if there is an annotation which matches the history item in terms of layer and feature
-     *     value, then this annotation should be highlighted.</li>
+     * value, then this annotation should be highlighted.</li>
      * <li>if there is no matching annotation, then the text should be highlighted</li>
      * </ul>
      */
@@ -869,7 +859,7 @@ public class ActiveLearningSidebar
     {
         actionShowSelectedDocument(aTarget, aRecord.getSourceDocument(),
                 aRecord.getOffsetCharacterBegin(), aRecord.getOffsetCharacterEnd());
-        
+
         // Since we have switched documents above (if it was necessary), the editor CAS should
         // now point to the correct one
         JCas jCas = getJCasProvider().get();
@@ -882,20 +872,20 @@ public class ActiveLearningSidebar
         // ... otherwise highlight the text
         else {
             setHighlight(aRecord);
-            
+
             info("No annotation could be highlighted.");
             aTarget.addChildren(getPage(), IFeedback.class);
         }
     }
-    
+
     private Optional<AnnotationFS> getMatchingAnnotation(CAS aCas, LearningRecord aRecord)
     {
         Type type = CasUtil.getType(aCas, alStateModel.getObject().getLayer().getName());
         Feature feature = type.getFeatureByBaseName(aRecord.getAnnotationFeature().getName());
         return selectAt(aCas, type, aRecord.getOffsetCharacterBegin(),
-                aRecord.getOffsetCharacterEnd()).stream()
-                .filter(fs -> aRecord.getAnnotation().equals(fs.getFeatureValueAsString(feature)))
-                .findFirst();
+                aRecord.getOffsetCharacterEnd()).stream().filter(
+                        fs -> aRecord.getAnnotation().equals(fs.getFeatureValueAsString(feature)))
+                        .findFirst();
     }
 
     private List<AnnotationSuggestion> getMatchingSuggestion(List<SuggestionGroup> aSuggestions,
@@ -920,15 +910,13 @@ public class ActiveLearningSidebar
             String aLabel)
     {
         return aSuggestions.stream()
-                .filter(group -> 
-                        aDocumentName.equals(group.getDocumentName()) &&
-                        aLayerId == group.getLayerId() &&
-                        (aFeature == null || aFeature == group.getFeature()) &&
-                        (aBegin == -1 || aBegin == group.getOffset().getBegin()) &&
-                        (aEnd == -1 || aEnd == group.getOffset().getEnd()))
+                .filter(group -> aDocumentName.equals(group.getDocumentName())
+                        && aLayerId == group.getLayerId()
+                        && (aFeature == null || aFeature == group.getFeature())
+                        && (aBegin == -1 || aBegin == group.getOffset().getBegin())
+                        && (aEnd == -1 || aEnd == group.getOffset().getEnd()))
                 .flatMap(group -> group.stream())
-                .filter(suggestion ->
-                        aLabel == null || aLabel.equals(suggestion.getLabel()))
+                .filter(suggestion -> aLabel == null || aLabel.equals(suggestion.getLabel()))
                 .collect(toList());
     }
 
@@ -942,18 +930,18 @@ public class ActiveLearningSidebar
         throws IOException
     {
         aTarget.add(mainContainer);
-        
+
         ActiveLearningUserState alState = alStateModel.getObject();
-        
+
         annotationPage.actionRefreshDocument(aTarget);
         learningRecordService.delete(aRecord);
-        
+
         // The history records caused suggestions to disappear. Since visibility is only fully
         // recalculated when new predictions come in, we need to update the visibility explicitly
         // here
         getMatchingSuggestion(alState.getSuggestions(), aRecord)
                 .forEach(suggestion -> suggestion.show(FLAG_SKIPPED | FLAG_REJECTED));
-        
+
         // Force the learning records model to be refreshed during rendering, showing the latest
         // state from the DB
         learningRecords.detach();
@@ -976,7 +964,7 @@ public class ActiveLearningSidebar
                 confirmationDialog.setConfirmAction(t -> deleteAnnotationByHistory(t, aRecord));
             }
         }
-        
+
         // If there is currently no suggestion (i.e. we ran out of suggestions before) there is a
         // good chance that deleting the history item makes suggestions become available again, so
         // we try to find a new one.
@@ -990,7 +978,7 @@ public class ActiveLearningSidebar
         throws IOException, AnnotationException
     {
         AnnotatorState state = getModelObject();
-        
+
         JCas jCas = this.getJCasProvider().get();
         Optional<AnnotationFS> anno = getMatchingAnnotation(jCas.getCas(), aRecord);
         if (anno.isPresent()) {
@@ -1002,8 +990,8 @@ public class ActiveLearningSidebar
 
     /**
      * Listens to the user setting a feature on an annotation in the main annotation editor. Mind
-     * that we do not need to listen to the creation of annotations since they have no effect on 
-     * the active learning sidebar as long as they have no features set.
+     * that we do not need to listen to the creation of annotations since they have no effect on the
+     * active learning sidebar as long as they have no features set.
      */
     @OnEvent
     public void onFeatureValueUpdated(FeatureValueUpdatedEvent aEvent)
@@ -1013,19 +1001,17 @@ public class ActiveLearningSidebar
 
         // If the user creates a new annotation at the site of the suggestion that is currently
         // offered to the user, then the AL should move on to the next available suggestion
-        if (
-                alState.isSessionActive() &&
-                (!alState.getSuggestion().isPresent() || (
-                aEvent.getUser().equals(state.getUser().getUsername()) &&
-                aEvent.getDocument().equals(state.getDocument()) &&
-                aEvent.getFeature().getLayer().equals(alState.getLayer()) &&
-                aEvent.getFeature().getName().equals(alState.getSuggestion().get().getFeature())))
-        ) {
+        if (alState.isSessionActive() && (!alState.getSuggestion().isPresent()
+                || (aEvent.getUser().equals(state.getUser().getUsername())
+                        && aEvent.getDocument().equals(state.getDocument())
+                        && aEvent.getFeature().getLayer().equals(alState.getLayer())
+                        && aEvent.getFeature().getName()
+                                .equals(alState.getSuggestion().get().getFeature())))) {
             reactToAnnotationsBeingCreatedOrDeleted(aEvent.getRequestTarget(),
                     aEvent.getFeature().getLayer());
         }
     }
-    
+
     /**
      * Listens to the user deleting an annotation in the main annotation editor.
      */
@@ -1037,34 +1023,31 @@ public class ActiveLearningSidebar
 
         // If the user creates a new annotation at the site of the suggestion that is currently
         // offered to the user, then the AL should move on to the next available suggestion
-        if (
-                alState.isSessionActive() &&
-                (!alState.getSuggestion().isPresent() || (
-                aEvent.getUser().equals(state.getUser().getUsername()) &&
-                aEvent.getDocument().equals(state.getDocument()) &&
-                aEvent.getLayer().equals(alState.getLayer())))
-        ) {
+        if (alState.isSessionActive() && (!alState.getSuggestion().isPresent()
+                || (aEvent.getUser().equals(state.getUser().getUsername())
+                        && aEvent.getDocument().equals(state.getDocument())
+                        && aEvent.getLayer().equals(alState.getLayer())))) {
             reactToAnnotationsBeingCreatedOrDeleted(aEvent.getRequestTarget(), aEvent.getLayer());
         }
     }
-    
+
     private void reactToAnnotationsBeingCreatedOrDeleted(IPartialPageRequestHandler aTarget,
             AnnotationLayer aLayer)
     {
         try {
             AnnotatorState state = getModelObject();
             ActiveLearningUserState alState = alStateModel.getObject();
-            
-//            // Make sure we know about the current suggestions and their visibility state
-//            refreshSuggestions();
-    
-            // Update visibility in case the annotation where the feature was set overlaps with 
+
+            // // Make sure we know about the current suggestions and their visibility state
+            // refreshSuggestions();
+
+            // Update visibility in case the annotation where the feature was set overlaps with
             // any suggestions that need to be hidden now.
-            PredictionTask.calculateVisibility(learningRecordService, annotationService,
+            PredictionUtil.calculateVisibility(learningRecordService, annotationService,
                     getAnnotationPage().getEditorCas().getCas(), state.getUser().getUsername(),
                     aLayer, alState.getSuggestions(), state.getWindowBeginOffset(),
                     state.getWindowEndOffset());
-    
+
             // Update the suggestion in the AL sidebar, but do not jump or touch the right
             // sidebar such that the user can happily continue to edit the annotation
             moveToNextRecommendation(aTarget, true);
@@ -1085,11 +1068,9 @@ public class ActiveLearningSidebar
         Predictions model = recommendationService.getPredictions(annotatorState.getUser(),
                 annotatorState.getProject());
 
-        if (
-                alStateModel.getObject().isSessionActive() && 
-                eventState.getUser().equals(annotatorState.getUser()) && 
-                eventState.getProject().equals(annotatorState.getProject())
-        ) {
+        if (alStateModel.getObject().isSessionActive()
+                && eventState.getUser().equals(annotatorState.getUser())
+                && eventState.getProject().equals(annotatorState.getProject())) {
             SourceDocument document = eventState.getDocument();
             VID vid = aEvent.getVid();
             Optional<AnnotationSuggestion> prediction = model.getPredictionByVID(document, vid);
@@ -1102,22 +1083,20 @@ public class ActiveLearningSidebar
 
             AnnotationSuggestion rejectedRecommendation = prediction.get();
             applicationEventPublisherHolder.get().publishEvent(
-                new ActiveLearningRecommendationEvent(this, eventState.getDocument(),
-                    rejectedRecommendation, annotatorState.getUser().getUsername(),
-                    eventState.getSelectedAnnotationLayer(), rejectedRecommendation.getFeature(),
-                    REJECTED, model.getPredictionsByTokenAndFeature(
-                    rejectedRecommendation.getDocumentName(),
-                    eventState.getSelectedAnnotationLayer(),
-                    rejectedRecommendation.getBegin(),
-                    rejectedRecommendation.getEnd(),
-                    rejectedRecommendation.getFeature())));
+                    new ActiveLearningRecommendationEvent(this, eventState.getDocument(),
+                            rejectedRecommendation, annotatorState.getUser().getUsername(),
+                            eventState.getSelectedAnnotationLayer(),
+                            rejectedRecommendation.getFeature(), REJECTED,
+                            model.getPredictionsByTokenAndFeature(
+                                    rejectedRecommendation.getDocumentName(),
+                                    eventState.getSelectedAnnotationLayer(),
+                                    rejectedRecommendation.getBegin(),
+                                    rejectedRecommendation.getEnd(),
+                                    rejectedRecommendation.getFeature())));
 
-            if (
-                    document.equals(annotatorState.getDocument()) && 
-                    vid.getLayerId() == alStateModel.getObject().getLayer().getId() && 
-                    prediction.get().equals(
-                            alStateModel.getObject().getSuggestion().orElse(null))
-            ) {
+            if (document.equals(annotatorState.getDocument())
+                    && vid.getLayerId() == alStateModel.getObject().getLayer().getId() && prediction
+                            .get().equals(alStateModel.getObject().getSuggestion().orElse(null))) {
                 moveToNextRecommendation(aEvent.getTarget(), false);
             }
             aEvent.getTarget().add(mainContainer);
@@ -1136,7 +1115,7 @@ public class ActiveLearningSidebar
         AnnotatorState eventState = aEvent.getAnnotatorState();
         SourceDocument document = state.getDocument();
         VID vid = aEvent.getVid();
-        
+
         Optional<AnnotationSuggestion> oRecommendation = model.getPredictionByVID(document, vid);
         if (!oRecommendation.isPresent()) {
             LOG.error("Could not find prediction in [{}] with id [{}]", document, vid);
@@ -1146,33 +1125,24 @@ public class ActiveLearningSidebar
         }
 
         AnnotationSuggestion acceptedSuggestion = oRecommendation.get();
-        
-        applicationEventPublisherHolder.get().publishEvent(
-            new ActiveLearningRecommendationEvent(this, eventState.getDocument(),
-                acceptedSuggestion, state.getUser().getUsername(),
-                eventState.getSelectedAnnotationLayer(), acceptedSuggestion.getFeature(),
-                ACCEPTED, model.getPredictionsByTokenAndFeature(
-                acceptedSuggestion.getDocumentName(),
-                eventState.getSelectedAnnotationLayer(),
-                acceptedSuggestion.getBegin(),
-                acceptedSuggestion.getEnd(),
-                acceptedSuggestion.getFeature())));
+
+        applicationEventPublisherHolder.get().publishEvent(new ActiveLearningRecommendationEvent(
+                this, eventState.getDocument(), acceptedSuggestion, state.getUser().getUsername(),
+                eventState.getSelectedAnnotationLayer(), acceptedSuggestion.getFeature(), ACCEPTED,
+                model.getPredictionsByTokenAndFeature(acceptedSuggestion.getDocumentName(),
+                        eventState.getSelectedAnnotationLayer(), acceptedSuggestion.getBegin(),
+                        acceptedSuggestion.getEnd(), acceptedSuggestion.getFeature())));
 
         // If the annotation that the user accepted is the one that is currently displayed in
         // the annotation sidebar, then we have to go and pick a new one
         ActiveLearningUserState alState = alStateModel.getObject();
-        if (
-                alState.isSessionActive() && 
-                alState.getSuggestion().isPresent() && 
-                eventState.getUser().equals(state.getUser()) && 
-                eventState.getProject().equals(state.getProject())
-        ) {
+        if (alState.isSessionActive() && alState.getSuggestion().isPresent()
+                && eventState.getUser().equals(state.getUser())
+                && eventState.getProject().equals(state.getProject())) {
             AnnotationSuggestion suggestion = alState.getSuggestion().get();
-            if (
-                    acceptedSuggestion.getOffset().equals(suggestion.getOffset()) && 
-                    vid.getLayerId() == suggestion.getLayerId() && 
-                    acceptedSuggestion.getFeature().equals(suggestion.getFeature())
-            ) {
+            if (acceptedSuggestion.getOffset().equals(suggestion.getOffset())
+                    && vid.getLayerId() == suggestion.getLayerId()
+                    && acceptedSuggestion.getFeature().equals(suggestion.getFeature())) {
                 moveToNextRecommendation(aEvent.getTarget(), false);
             }
             aEvent.getTarget().add(mainContainer);
@@ -1184,7 +1154,7 @@ public class ActiveLearningSidebar
     {
         renderHighlights(aEvent.getVDocument());
     }
-    
+
     private void renderHighlights(VDocument aVDoc)
     {
         LOG.trace("Active learning sidebar rendering highlights");
@@ -1204,13 +1174,12 @@ public class ActiveLearningSidebar
                     highlightDocumentName, currentDoc);
             return;
         }
-         
+
         if (highlightVID != null) {
             aVDoc.add(new VAnnotationMarker(this, VMarker.FOCUS, highlightVID));
         }
         else {
-            LOG.trace(
-                    "Active learning sidebar annotation highlight is not set");
+            LOG.trace("Active learning sidebar annotation highlight is not set");
         }
 
         if (highlightSpan != null) {
@@ -1237,24 +1206,24 @@ public class ActiveLearningSidebar
         alState.setSuggestions(
                 activeLearningService.getSuggestions(state.getUser(), alState.getLayer()));
     }
-    
+
     @OnEvent
     public void onDocumentReset(AfterDocumentResetEvent aEvent)
     {
         reactToChangeInPredictions(aEvent.getRequestTarget());
     }
-    
+
     @OnEvent
     public void onPredictionsSwitched(AjaxPredictionsSwitchedEvent aEvent)
     {
         reactToChangeInPredictions(aEvent.getTarget());
-        // As a reaction to the change in predictions, the highlights may have to be placed at 
+        // As a reaction to the change in predictions, the highlights may have to be placed at
         // a different location. This the prediction switch is announced late in the rendering
-        // process and the highlights have already been added to the VDocument at this time, 
+        // process and the highlights have already been added to the VDocument at this time,
         // we need to remove and re-add them. renderHighlights() takes care of this.
         renderHighlights(aEvent.getVDocument());
     }
-    
+
     private void reactToChangeInPredictions(IPartialPageRequestHandler aTarget)
     {
         ActiveLearningUserState alState = alStateModel.getObject();
@@ -1268,17 +1237,17 @@ public class ActiveLearningSidebar
 
         // Make sure we know about the current suggestions and their visibility state
         refreshSuggestions();
-        
+
         // If there is currently a suggestion displayed in the sidebar, we need to check if it
         // is still relevant - if yes, we need to replace it with its current counterpart since.
-        // if no counterpart exists in the current suggestions, then we need to load a 
+        // if no counterpart exists in the current suggestions, then we need to load a
         // suggestion from the current list.
         if (alState.getSuggestion().isPresent()) {
             AnnotationSuggestion activeSuggestion = alState.getSuggestion().get();
             // Find the groups which matches the active recommendation
             Optional<AnnotationSuggestion> updatedSuggestion = getMatchingSuggestion(
                     alState.getSuggestions(), activeSuggestion).stream().findFirst();
-            
+
             if (updatedSuggestion.isPresent()) {
                 LOG.debug("Replacing outdated suggestion {} with new suggestion {}",
                         alState.getCurrentDifference().get().getFirst().getId(),
@@ -1289,7 +1258,7 @@ public class ActiveLearningSidebar
                     highlightVID = updatedSuggestion.get().getVID();
                 }
 
-                // We found a matching suggestion, but we look for its second-best. So for the 
+                // We found a matching suggestion, but we look for its second-best. So for the
                 // moment we assume that the second-best has not changed and we simply fake
                 // a delta
                 Delta fakeDelta = new Delta(updatedSuggestion.get(),
